% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt_table.r
\name{ctt_table}
\alias{ctt_table}
\title{ctt_table() produces a table with classical test theory estimates.}
\usage{
ctt_table(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
data frame with CTT estimates, generated with CTT::itemAnalysis()
}
\description{
ctt_table() produces a table with classical test theory estimates.
}
\examples{

ctt_table(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
