% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_grm2_align_wjs_wlsmv.r
\name{fit_grm2_align_wjs_wlsmv}
\alias{fit_grm2_align_wjs_wlsmv}
\title{fit_grm2_align_wjs_wlsmv() it fits a between country GRM with alignment method,
with a complex sample design using MPLUS and MplusAutomation.
The "\emph{wjs}" version uses taylor series linearization, using strafication, clusters and weights.}
\usage{
fit_grm2_align_wjs_wlsmv(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
fit_grm2_align_wjs_wlsmv() it fits a between country GRM with alignment method,
with a complex sample design using MPLUS and MplusAutomation.
The "\emph{wjs}" version uses taylor series linearization, using strafication, clusters and weights.
}
\examples{

scale_001_align <- fit_grm2_align_wlsmv(
scale_info = scales_data,
scale_num  = 1,
data = data_model)


}
