% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_grm2_inv2.r
\name{fit_grm2_inv2}
\alias{fit_grm2_inv2}
\title{fit_grm2_inv1() it fits GRM-metric between countries using MPLUS and MplusAutomation}
\usage{
fit_grm2_inv2(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
fit_grm2_inv1() it fits GRM-metric between countries using MPLUS and MplusAutomation
}
\examples{

scale_001_inv2 <- fit_grm2_inv2(
scale_info = scales_data,
scale_num  = 1,
data = data_model)


}
