% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_per_variable.r
\name{frequency_per_variable}
\alias{frequency_per_variable}
\title{frequency_per_variable() calculates proportions of category use, from a polytomous scale}
\usage{
frequency_per_variable(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
data frame with proportions and histograms, using skimr::inline_hist
}
\description{
frequency_per_variable() calculates proportions of category use, from a polytomous scale
}
\examples{

frequency_per_variable(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
