% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descriptives.r
\name{get_descriptives}
\alias{get_descriptives}
\title{get_descriptives() produces a descriptive table from a response data frame}
\usage{
get_descriptives(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
data frame with descriptives estimates using libraries such as stats, moments, and skimr
}
\description{
get_descriptives() produces a descriptive table from a response data frame
}
\examples{

get_descriptives(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
