% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_empty.r
\name{get_empty}
\alias{get_empty}
\title{get_empty() estimates the amount of cases with empty rows}
\usage{
get_empty(data)
}
\arguments{
\item{data}{a data frame responses, where rows = observations, and columns = variables}
}
\value{
a table of missing, distinguishing missing by
}
\description{
get_empty() estimates the amount of cases with empty rows
}
\examples{

data_with_countries \%>\%
dplyr::select(countries, item_001:item_050) \%>\%
split(.$countries) \%>\%
purrr::map( ~ get_empty(.)) \%>\%
dplyr::bind_rows(., .id = 'IDCNTRY')


}
