% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit.r
\name{get_fit}
\alias{get_fit}
\title{get_fit() produces a table displaying obtained fit statistics}
\usage{
get_fit(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a table
}
\description{
get_fit() produces a table displaying obtained fit statistics
}
\examples{

get_fit(scale_001)


}
