% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inv_fit.r
\name{get_inv_fit}
\alias{get_inv_fit}
\title{get_inv_fit() retrieves model fit indexes from GRM models to generate the invariance table}
\usage{
get_inv_fit(model, model_name)
}
\arguments{
\item{model}{a response model, with a MplusAutomation model object}

\item{model_name}{argument to define the fitted model (e.g., pooled, strict, scalar, configural, base)}
}
\description{
get_inv_fit() retrieves model fit indexes from GRM models to generate the invariance table
}
\examples{

fit_1 <- get_inv_fit(
         model = inv_1, 
         model_name = 'strict'
         )

}
