% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lambda.r
\name{get_lambda}
\alias{get_lambda}
\title{get_lambda() retrieves the factor loadings of a latent variable model}
\usage{
get_lambda(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
data frame with with descriptive values
}
\description{
get_lambda() retrieves the factor loadings of a latent variable model
}
\examples{

get_lambda(model)


}
