% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pattern.r
\name{get_pattern}
\alias{get_pattern}
\title{get_pattern() generates a string to represent a response pattern}
\usage{
get_pattern(data, variables)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{variables}{selected variables to produce a response pattern}
}
\value{
mplus_object generated with MplusAutomation
}
\description{
get_pattern() generates a string to represent a response pattern
}
\examples{

data_with_pattern <- get_pattern(
data = data_model,
variables  = c('f1','f2','f3','f4','f5','f6')
)


}
