% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_score.r
\name{get_score}
\alias{get_score}
\title{get_score() retrieves the theta score and re-scale these at mean zero, and sd one at the population level}
\usage{
get_score(model, name)
}
\arguments{
\item{model}{an mplus object generated with techr::fit_grm2 or techr::fit_pcm}

\item{name}{the name of the score included in the returned data frame}
}
\value{
data frame with id_i and a score vector with mean == 0 and sd == 1, at the population level
}
\description{
get_score() retrieves the theta score and re-scale these at mean zero, and sd one at the population level
}
\examples{

score_01 <- fit_grm2(
            scale_info = scales_data,
            scale_num  = 1,
            data = data_model) \%>\%
            get_score(., 'score')


}
