% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_theta.r
\name{get_theta}
\alias{get_theta}
\title{get_theta() retrieves the theta score generated with techr::fit_grm2 or techr::fit_pcm}
\usage{
get_theta(model, name)
}
\arguments{
\item{model}{an mplus object generated with techr::fit_grm2 or techr::fit_pcm}

\item{name}{the name of the score included in the returned data frame}
}
\value{
data frame with id_i and a score vector with the defined name
}
\description{
get_theta() retrieves the theta score generated with techr::fit_grm2 or techr::fit_pcm
}
\examples{

theta_01 <- fit_grm2(
            scale_info = scales_data,
            scale_num  = 1,
            data = data_model) \%>\%
            get_theta(., 'score')


}
