% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_summary.r
\name{missing_summary}
\alias{missing_summary}
\title{missing_summary() generates missing summary}
\usage{
missing_summary(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
a table of missing, distinguishing between complete response, partial, and fully missing
}
\description{
missing_summary() generates missing summary
}
\examples{

missing_summary(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
