% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_table.r
\name{missing_table}
\alias{missing_table}
\title{missing_table() generates missing summary}
\usage{
missing_table(data)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}
}
\value{
a table of missing, distinguishing between complete response, partial, and fully missing
}
\description{
missing_table() generates missing summary
}
\examples{

missing_table(
data = scale_data
) \%>\%
knitr::kabel(., digits = 2)


}
