% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_lubbe.r
\name{pa_lubbe}
\alias{pa_lubbe}
\title{pa_lubbe() calculates a parallel analysis using a response data frame}
\usage{
pa_lubbe(data, scale_num, scale_info, n_sim)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{n_sim}{specify the number of simulated correlations, n_sim = 1000 is recommended}
}
\value{
a data frame generated depicting observed and simulated eigenvalues, and number of factors
}
\description{
pa_lubbe() calculates a parallel analysis using a response data frame
}
\examples{

pa_lubbe(
data = scale_data,
scale_num = 1,
scale_info = items_data,
n_sim = 1000
) \%>\%
knitr::kabel(., digits = 2)


}
