% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_lubbe_plot.r
\name{pa_lubbe_plot}
\alias{pa_lubbe_plot}
\title{pa_lubbe_plot() plot the results of a parallel analysis generated with the pa_lubb() function}
\usage{
pa_lubbe_plot(data)
}
\arguments{
\item{data}{it uses an input a pa_lubb() results, UCP adjusted parallel analysis (Lubbe, 2019)}
}
\value{
a ggplot object
}
\description{
pa_lubbe_plot() plot the results of a parallel analysis generated with the pa_lubb() function
}
\examples{

pa_lubbe_plot(pa_lubbe_data)


}
