% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_error.r
\name{plot_error}
\alias{plot_error}
\title{plot_error() plot a scatter of errors and theta locations with their histograms, using library(ggplot2)}
\usage{
plot_error(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a ggplot object
}
\description{
plot_error() plot a scatter of errors and theta locations with their histograms, using library(ggplot2)
}
\examples{

plot_error(scale_001)


}
