% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infit.r
\name{plot_infit}
\alias{plot_infit}
\title{plot_infit() plot infit results from a Rasch model over a response data frame}
\usage{
plot_infit(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
a base plot
}
\description{
plot_infit() plot infit results from a Rasch model over a response data frame
}
\examples{

plot_infit(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
