% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_item_map_grm3.r
\name{plot_item_map_grm3}
\alias{plot_item_map_grm3}
\title{plot_item_map_grm3() generates an item person map using a GRM model with 4 response categories}
\usage{
plot_item_map_grm3(scale_info, scale_num, model)
}
\arguments{
\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a item map plot generated with library(ggplot2)
}
\description{
plot_item_map_grm3() generates an item person map using a GRM model with 4 response categories
}
\examples{

plot_item_map_grm3(scales_data, 2, scale_002)


}
