% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_pattern_plain.r
\name{plot_missing_pattern_plain}
\alias{plot_missing_pattern_plain}
\title{plot_missing_pattern_plain() generates a plot with missing patterns using VIM::aggr}
\usage{
plot_missing_pattern_plain(data, scale_num, scale_info, n_sample)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}

\item{n_sample}{define number of random sample}
}
\value{
a plot of missing patters generated with VIM::aggr
}
\description{
plot_missing_pattern_plain() generates a plot with missing patterns using VIM::aggr
}
\examples{

plot_missing_pattern(
data = data_responses,
scale_num = 1,
scale_info = scales_data
)

}
