% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_missing_summary.r
\name{plot_missing_summary}
\alias{plot_missing_summary}
\title{plot_missing_summary() it plots missing summary}
\source{
\url{https://www.r-bloggers.com/round-values-while-preserve-their-rounded-sum-in-r/} and
\url{http://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}
}
\usage{
plot_missing_summary(missing_summary)
}
\arguments{
\item{missing_summary}{a data frame generated with techr::missing_summary}
}
\value{
a plot generated with library(ggplot2) and library(waffle)
}
\description{
plot_missing_summary() it plots missing summary
}
\examples{

missing_summary(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
plot_missing_summary()


}
