% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theta_caterpillar.r
\name{plot_theta_caterpillar}
\alias{plot_theta_caterpillar}
\title{plot_theta_caterpillar() produces a caterpillar plot for theta scores}
\usage{
plot_theta_caterpillar(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a ggplot object
}
\description{
plot_theta_caterpillar() produces a caterpillar plot for theta scores
}
\examples{

plot_theta_caterpillar(scale_001)


}
