% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_alpha.r
\name{raw_alpha}
\alias{raw_alpha}
\title{raw_alpha() calculates the alpha coefficient from a response table.}
\usage{
raw_alpha(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
a numeric object, calculated with psych::alpha
}
\description{
raw_alpha() calculates the alpha coefficient from a response table.
}
\examples{

raw_alpha(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
print


}
