% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability_separation.r
\name{reliability_separation}
\alias{reliability_separation}
\title{reliability_separation() calculates separation reliability of theta scores realizations}
\usage{
reliability_separation(model)
}
\arguments{
\item{model}{it uses an input a mplus_object generated with MplusAutomation}
}
\value{
a numeric object
}
\description{
reliability_separation() calculates separation reliability of theta scores realizations
}
\examples{

reliability_separation(scale_001)


}
