% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_infit.r
\name{scale_infit}
\alias{scale_infit}
\title{scale_infit() generates a infit table results from a Rasch model over a response data frame}
\usage{
scale_infit(data, scale_num, scale_info)
}
\arguments{
\item{data}{a data frame, where rows = observations, and columns = variables}

\item{scale_num}{a number, that identifies a unique set of items within the scale_info table}

\item{scale_info}{a data frame where items memmbership to scale is uniquely identify}
}
\value{
a table of infit results, generated with TAM::tam.fit, showing only those satisfying the condition 'abs(Infit_t) >2 & Infit > 1.33 | abs(Infit_t) >2 & Infit <  .75)'
}
\description{
scale_infit() generates a infit table results from a Rasch model over a response data frame
}
\examples{

scale_infit(
data = scale_data,
scale_num = 1,
scale_info = items_data
) \%>\%
knitr::kabel(., digits = 2)


}
